requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
affine==2.4.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1696154234947/work
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
graphviz==0.20.3
grpcio @ file:///C:/b/abs_ccut8qek12/croot/grpc-suite_1681912611694/work
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.1
importlib-metadata @ file:///C:/b/abs_c1egths604/croot/importlib_metadata-suite_1704813568388/work
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
labelme==3.16.7
lazy_loader==0.4
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.5.0
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.1
notebook_shim==0.2.4
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
open3d==0.18.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.0
pandas==1.3.4
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
pip-autoremove==0.10.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.3.10
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools-scm==8.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
tenacity==8.5.0
tensorboard==1.15.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
torchinfo @ file:///home/conda/feedstock_root/build_artifacts/torchinfo_1684097133203/work
torchsummary==1.5.1
torchvision==0.11.2
torchviz==0.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///C:/b/abs_72cdotwc_6/croot/typing_extensions_1705599364138/work
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug @ file:///C:/b/abs_8bittcw9jr/croot/werkzeug_1716533366070/work
widgetsnbextension==4.0.11
zipp==3.19.2
